Modify fixed-point to allow both 16-bit and 32-bit types, and consolidate ordinary fixed-point representations with accumulated ones.
darcs-hash:20071024025427-d1b56-5a27f4d8c9c606ca5517abc5018321a282e1377f.gz

+++ d2/pixel_accum.h
@@ -137,7 +137,7 @@ inline pixel_accum operator*(double d, const pixel_accum &p) {
 }
 
 inline std::ostream &operator<<(std::ostream &o, const pixel_accum &p) {
-	o << "[" << p[0] << " " << p[1] << " " << p[2] << "]";
+	o << "[" << (double) p[0] << " " << (double) p[1] << " " << (double) p[2] << "]";
 	return o;
 }
